# platform/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Codex Control Platform API
  version: 0.1.0
  description: Account, desktop enrollment, Host allocation, and localhost V3 proxy control plane. The production outbound Relay transport is a later deployment phase.
servers:
  - url: http://localhost:8088/api/v1
    description: Local Compose stack
...
paths:
...
  /readiness:
    get:
      tags: [Platform]
      operationId: readiness
      responses:
        '200': { description: Database is reachable }
        '503': { description: Database is unavailable }
...
  /auth/register:
    post:
      tags: [Authentication]
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Account and session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/login:
    post:
      tags: [Authentication]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/refresh:
    post:
      tags: [Authentication]
      operationId: refreshSession
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string, description: Required for a desktop client; browser uses HttpOnly cookie }
                clientType: { type: string, enum: [browser, desktop], default: browser }
      responses:
        '200':
          description: Rotated session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/session:
    get:
      tags: [Authentication]
      security: [{ accountSession: [] }]
      operationId: currentSession
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Authentication]
      security: [{ accountSession: [] }]
      operationId: logout
      responses:
        '204': { description: Session revoked and cookies cleared }
        '403': { $ref: '#/components/responses/CsrfRejected' }
...
components:
  securitySchemes:
    accountSession:
      type: apiKey
      in: cookie
      name: ccc_platform_access
      description: Browser may use HttpOnly cookie; native clients use the same opaque value as a Bearer token.
...
  responses:
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Resource already exists or is not in the required state
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    CsrfRejected:
      description: Browser CSRF token is absent or invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Sensitive endpoint attempt limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, format: password, minLength: 12, maxLength: 72 }
        displayName: { type: string, maxLength: 80, description: Optional; defaults to the email local part }
        clientType: { type: string, enum: [browser, desktop], default: browser }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
        clientType: { type: string, enum: [browser, desktop], default: browser }
    User:
      type: object
      required: [id, email, displayName, status, emailVerified, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        status: { type: string, enum: [active, pending, disabled] }
        emailVerified: { type: boolean }
        createdAt: { type: string, format: date-time }
    AuthResponse:
      type: object
      required: [user, accessExpiresAt, refreshExpiresAt, csrfToken]
      properties:
        user: { $ref: '#/components/schemas/User' }
        accessExpiresAt: { type: string, format: date-time }
        refreshExpiresAt: { type: string, format: date-time }
        csrfToken: { type: string }
        accessToken: { type: string, description: Returned only to desktop client type }
        refreshToken: { type: string, description: Returned only to desktop client type }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, timestamp]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: [string, 'null'] }
            timestamp: { type: string, format: date-time }
            details: { type: object, additionalProperties: true }